+++ cmd/extend.go
@@ -6,8 +6,8 @@ import (
 	"os"
 	"time"
 
-	"azpim/internal/auth"
-	"azpim/internal/pim"
+	"github.com/bgs113/azpim/internal/auth"
+	"github.com/bgs113/azpim/internal/pim"
 
 	"github.com/spf13/cobra"
 )